pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["fniad <artkamproject@gmail.com>"]
readme = "README.md"
packages = [{include = "drf_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
psycopg2-binary = "^2.9.9"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
python-decouple = "^3.8"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
requests = "^2.31.0"
stripe = "^7.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
docker = "^7.0.0"
django-celery-beat = "^2.5.0"
python-crontab = "^3.0.0"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
